signaling client file name changed
